+++ tests/conftest.py
@@ -14,7 +14,7 @@ import os
 
 import earthaccess
 import pytest
-from harmony import Client, Environment, Request
+from harmony import CapabilitiesRequest, Client, Environment, Request
 
 environment_mapping = {
     'production': Environment.PROD,
@@ -37,8 +37,15 @@ class AutotesterRequest(Request):
         super().__init__(**kwargs, labels=labels)
 
 
+service_collections = json.loads(os.environ.get('SERVICE_COLLECTIONS', '[]'))
+
+
 @pytest.fixture(
-    params=json.loads(os.environ.get('SERVICE_COLLECTIONS', '[]')),
+    params=service_collections,
+    ids=[
+        f'{collection["short_name"]}-{collection["version"]}-{collection["concept_id"]}'
+        for collection in service_collections
+    ],
     scope='session',
 )
 def service_collection(request):
@@ -68,9 +75,16 @@ def earthaccess_login():
 
 
 @pytest.fixture(scope='session')
-def test_output_file():
-    """The path to where the failed test information should be written."""
-    test_directory = os.environ.get('TEST_DIRECTORY')
+def test_output_file(request):
+    """The path to where the failed test information should be written.
+
+    Defaults to the directory of the tests being run if `TEST_DIRECTORY` is
+    not set (local development).
+
+    """
+    test_directory = os.environ.get('TEST_DIRECTORY') or (
+        request.session.items[0].path.parent
+    )
     return f'{test_directory}/test_output.json'
 
 
@@ -81,3 +95,16 @@ def failed_tests(test_output_file):
     yield failed_test_information
     with open(test_output_file, 'w', encoding='utf-8') as file_handler:
         json.dump(failed_test_information, file_handler, indent=2)
+
+
+def get_configured_variable_names(
+    harmony_client: Client, collection_id: str
+) -> list[str]:
+    """Get the name of the configured variables for the collection.
+
+    Calls Harmony's capabilities endpont for the collection and returns a list
+    of all configured variable names.
+    """
+    cap_request = CapabilitiesRequest(collection_id=collection_id)
+    capabilities = harmony_client.submit(cap_request)
+    return [v['name'] for v in capabilities.get('variables')]